--- src/auth/jwt-auth.guard.ts
@@ -1,5 +1,0 @@
-import { Injectable } from "@nestjs/common";
-import { AuthGuard } from "@nestjs/passport";
-
-@Injectable()
-export class JwtAuthGuard extends AuthGuard('jwt') {}
