--- src/health-check/health-check/dto/create-health-check.dto.ts
@@ -1,1 +1,0 @@
-export class CreateHealthCheckDto {}
