--- src/health-check/health-check/dto/update-health-check.dto.ts
@@ -1,4 +1,0 @@
-import { PartialType } from '@nestjs/mapped-types';
-import { CreateHealthCheckDto } from './create-health-check.dto';
-
-export class UpdateHealthCheckDto extends PartialType(CreateHealthCheckDto) {}
