--- src/health-check/health-check/entities/health-check.entity.ts
@@ -1,1 +1,0 @@
-export class HealthCheck {}
