+++ src/health-check/schemas/health-check.schema.ts
@@ -1,9 +1,9 @@
 import { Prop, Schema, SchemaFactory } from "@nestjs/mongoose";
 import { Pet } from "src/pet/schemas/pet.schema";
-import { HealthStatus } from "../../enums/health-status.enum";
+import { HealthStatus } from "../enums/health-status.enum";
 import mongoose, { Document } from "mongoose";
 import { User } from "src/auth/schemas/user.schema";
-import { CheckingType } from "../../enums/checking-type.enum";
+import { CheckingType } from "../enums/checking-type.enum";
 
 export type HealthCheckDocument = HealthCheck & Document;
 @Schema({
